Add isFinite() methods in master branch.
Just in case someone is using these methods in Java7 branch, then they
update to master branch to use Java8, they shouldn't get a build error.
The methods are annotated as deprecated, so the user will still get a
build warning, but at least it will tell them how to fix it.